Remove click listeners on editor frame
These click handlers were there to show/hide the editor selection menu,
but that's handled by clicks on the AddButtonContainer. The onClick
handler was also calling 'preventDefault()' on the events, which was
causing the browser to ignore the default behavior of clicks on the
file input, which is used to trigger the file manager. Since we don't
even need the click handlers attached to the frame, just remove them
since that will resolve the issue of the file manager not opening.